Implements TDAmeritrade auth + data
New testing fixtures for oauth handshake, includes a feature flag
to remove the monkey patching of td_token.json, and instead relies
on a valid token in that file. This helps VCR record live API calls and
persist as a cassette for future testing.

Also takes care of a few dependency issues and comment cleanup

Inject td client where needed